api/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
pip==22.0.2
setuptools==59.6.0
pytz==2022.5
sqlparse==0.4.3
psycopg2>=2.8
